sched/tune: Rename dynamic_boost parameter to sched_boost
This was confusing to deal with given that it had the same name as the
Dynamic Schedtune Boost framework. It will be more apt to call it
sched_boost given that it was created to work with the sched_boost
feature in CAF devices.

The new tunable can be found in /dev/stune/*/schedtune.sched_boost

Change-Id: Iab528ade86aafd494d8186b0dc547ce190b460a6
Signed-off-by: joshuous <joshuous@gmail.com>
Signed-off-by: kdrag0n <dragon@khronodragon.com>