make: Add make wrapper for builds
This makes it possible to simply press `make` and have a kernel. It will
download a toolchain, create a configuration, and so forth. The .config
file will only be created if it does not already exist, making it
possible to edit out/.config manually during development. This takes
care of setting all relevant environment variables, and final kernels
will wind up in out/arch/arm64/boot/Image.gz-dtb.

Change-Id: I20058df48fd90df4934f4a1281f8d175e932edb2
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Signed-off-by: YumeMichi <do4suki@gmail.com>