initramfs: Expose in_recovery boolean variable
On A/B devices, skip_initramfs is present during all normal boots and
omitted by the bootloader only when booting into recovery. Leverage this
fact to determine whether recovery is being booted on the kernel level.
This creates an opportunity for kernel drivers to detect such scenarios
and react as appropriate to ensure an optimal experience in custom
recoveries such as TWRP.

Change-Id: I77b4ec92a2b0d98ecac98e1fc582ad6b327b2796
Signed-off-by: kdrag0n <dragon@khronodragon.com>