f2fs/rapid_gc: Do not wait until all invalid blocks are discarded
Undiscarded blocks may not reach 0 for various reasons, such as
discard granularity settings and discard thread terminating before GC ends,
and when that happens, GC thread enters an indefinite loop.

Remove this logic entirely.

This is not much of an issue since discard thread does not rely on rapid GC.

Change-Id: If9180610969f777c2e8348f1e1d27debe3be9fa3
Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>
Signed-off-by: kdrag0n <dragon@khronodragon.com>