cpu_input_boost: Allow max boost when the screen is off
In some cases, we may actually *want* to boost while the screen is off.
Drivers that use max boost should check if the screen is on depending on
their use of boosting.

One example is the early wake boost kick in drm_bridge, which is blocked
when these checks are in place.

Change-Id: I2dd238c7f029ea1b30354b36c026fb87f441f9b8
Signed-off-by: Danny Lin <danny@kdrag0n.dev>