kthread: Add API to create kthreads bound to low-power CPUs
Some unimportant kthreads do not need high-performance and can safely be
bound to low-power CPUs for energy efficiency.

This commit adds an API to create kthreads bound to low-power CPUs. This
affinity cannot be altered by the kernel or userspace.

This extends the performance-critical kthread support introduced in
commit c4a0ef7e1d1e ("kernel: Add API to mark IRQs and kthreads as
performance critical").

Change-Id: I9b5373221bd57029fb1687d1a7cdf5d8bece9350
Signed-off-by: Danny Lin <danny@kdrag0n.dev>