clk: qcom: clk-cpu-osm: Query current CPU frequency from OSM
Instead of getting the index and consulting the in-kernel frequency
table for the frequency corresponding to the fetched index, read the XO
clock multiplier for the current frequency directly from OSM and
multiply it by the XO rate. This ensures that if the kernel frequency
table has been altered, accurate readings can still be obtained.

300 MHz (OSM_INIT_RATE) is handled separately due to the use of a
separate clock source. All other frequencies are derived from the 19.2
MHz XO clock.

Change-Id: Iff6ca9c9e61c7179299f5c6de32556d7530606d8
Signed-off-by: Danny Lin <danny@kdrag0n.dev>