Revert "sched/walt: Reduce window size to 8 ms and load history to 64 ms"
This reverts commit 510d7e05631cef5901d7ef2c0209e4ed865cb260.

This causes the big cluster to use much higher frequencies than
necessary, often ramping to 2323200 KHz as soon as input is received.
The frequency also often hovers around 1689600 KHz when the screen is
off. There aren't any noticeable performance improvements either when
SchedTune boosting is used, so revert to the default 20 ms window and
100 ms load history.

Test: big cluster freqs during screen-off and screen-on return to normal
Change-Id: Id761e93b76b1c908a8f041f34d51f11ce5b2308f
Signed-off-by: Danny Lin <danny@kdrag0n.dev>