cpu_input_boost: Track input event times and expose timeout helper
Sometimes we may want to limit certain boost kicks to only take effect
within a certain timeout after receiving the last input event to save
power. For example, we may want to boost when an intensive process
forks to improve fluidity, but boosting when the fork was not initiated
by input is wasteful.

Change-Id: I2db24309890c727e4b0660771afaf01e94f43380
Signed-off-by: Danny Lin <danny@kdrag0n.dev>