cpu_input_boost: Add early wake boost on power button press
Boost the CPU preemptively upon a power button press to improve wake-up
speeds. The power button being pressed does not always translate to a
wake, but the power cost from a such a brief boost should not be a
deal-breaker.

In testing, the device woke up 10 ms faster on average after this
change.

Change-Id: I693aaf69aba9e5f4f00edc1df78621028bc06a4c
Signed-off-by: Danny Lin <danny@kdrag0n.dev>