exec: Fix potential race condition in boost task checker
A race condition where the task's comm is updated while it is being
compared may occur because we check the comm directly. Fix this by using
the get_task_comm helper which copies the comm for comparison.

Additionally, move and rename the checker function to reflect the change
in arguments.

Change-Id: I13015a0e99ea6eba0f13cf156790fb8d58ad6abf
Signed-off-by: Danny Lin <danny@kdrag0n.dev>