sched/fair.c: Don't allow SchedTune boosted tasks to be migrated to small cores
We want boosted tasks to run on big cores. But CAF's load balancer
changes do not account for SchedTune boosting, so this allows for
boosted tasks to be migrated to a suboptimal core. Let's mitigate
this by setting the LBF_IGNORE_BIG_TASKS for tasks migrating from a
larger capacity core to a smaller one and to check if the task is
SchedTune boosted. If both are true, do not migrate this task.

Change-Id: I0fb284518c265575cd4af0d95eadae6b7a2c82e6
Signed-off-by: Zachariah Kennedy <zkennedy87@gmail.com>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>